Read lyrics from *.txt file
Before trying to read lyrics embedded in the file (MP3, OGG etc.), check if an external lyrics file has been provided.
If so, try to read that file instead. If no such file exists, or the reading fails; try the old methods.
The UTF-8 character encoding is enforced to support german letters.